Use built-in tuple debug formatter
This replaces the clever and custom-made debug impl of gzip::Flags with
one that uses Rust's built-in [`DebugTuple`] formatter. The advantages
are that the code is less clever, may seem more familiar to Rust devs,
and that is will automatically support 'alternate mode' (`{:#?}`,
breaking up the fields over multiple lines). The disadvantages are that
we have to resort to a weird `&format_args!("..")` dance (because the
param given to `.field` needs to be `&dyn Debug`) and that the alternate
mode support doesn't seem very useful here.

[`DebugTuple`]: https://doc.rust-lang.org/1.36.0/std/fmt/struct.DebugTuple.html